# Introduce SortedLabelValues/LabelValues to speedup queries for high cardinality