If a label has a lot of values, sorting it after checking the matcher can save a good chunk of query time.