@brian-brazil 2020-06-24
This seems fine. What do the benchmarks show?