@codesome 2020-06-24
> What do the benchmarks show?

Haven't run any benchmarks yet. We detected this via CPU profiles. I will post the benchmark results after I run them.