@codesome 2020-06-24
Run 1:
```
benchmark                                                               old ns/op     new ns/op     delta
BenchmarkPostingsForMatchers/Head/n="1"-8                               857           873           +1.87%
BenchmarkPostingsForMatchers/Head/n="1",j="foo"-8                       1118          1143          +2.24%
BenchmarkPostingsForMatchers/Head/j="foo",n="1"-8                       1097          1174          +7.02%
BenchmarkPostingsForMatchers/Head/n="1",j!="foo"-8                      1336          1350          +1.05%
BenchmarkPostingsForMatchers/Head/i=~".*"-8                             162388278     99264870      -38.87%
BenchmarkPostingsForMatchers/Head/i=~".+"-8                             193966535     199994259     +3.11%
BenchmarkPostingsForMatchers/Head/i=~""-8                               115998729     119508872     +3.03%
BenchmarkPostingsForMatchers/Head/i!=""-8                               90338212      97925978      +8.40%
BenchmarkPostingsForMatchers/Head/n="1",i=~".*",j="foo"-8               161094227     100264573     -37.76%
BenchmarkPostingsForMatchers/Head/n="1",i=~".*",i!="2",j="foo"-8        162627302     101053412     -37.86%
BenchmarkPostingsForMatchers/Head/n="1",i!=""-8                         96382488      97642733      +1.31%
BenchmarkPostingsForMatchers/Head/n="1",i!="",j="foo"-8                 95588603      97039409      +1.52%
BenchmarkPostingsForMatchers/Head/n="1",i=~".+",j="foo"-8               192843307     201346606     +4.41%
BenchmarkPostingsForMatchers/Head/n="1",i=~"1.+",j="foo"-8              98043571      42998738      -56.14%
BenchmarkPostingsForMatchers/Head/n="1",i=~".+",i!="2",j="foo"-8        194609209     194450848     -0.08%
BenchmarkPostingsForMatchers/Head/n="1",i=~".+",i!~"2.*",j="foo"-8      290178661     243521697     -16.08%
BenchmarkPostingsForMatchers/Block/n="1"-8                              31767         31917         +0.47%
BenchmarkPostingsForMatchers/Block/n="1",j="foo"-8                      595909        534274        -10.34%
BenchmarkPostingsForMatchers/Block/j="foo",n="1"-8                      595858        545116        -8.52%
BenchmarkPostingsForMatchers/Block/n="1",j!="foo"-8                     597123        588109        -1.51%
BenchmarkPostingsForMatchers/Block/i=~".*"-8                            88874712      88604044      -0.30%
BenchmarkPostingsForMatchers/Block/i=~".+"-8                            106109029     125049364     +17.85%
BenchmarkPostingsForMatchers/Block/i=~""-8                              31964972      45528464      +42.43%
BenchmarkPostingsForMatchers/Block/i!=""-8                              24302758      37081367      +52.58%
BenchmarkPostingsForMatchers/Block/n="1",i=~".*",j="foo"-8              89734905      91631974      +2.11%
BenchmarkPostingsForMatchers/Block/n="1",i=~".*",i!="2",j="foo"-8       89186350      88838264      -0.39%
BenchmarkPostingsForMatchers/Block/n="1",i!=""-8                        23007373      37376000      +62.45%
BenchmarkPostingsForMatchers/Block/n="1",i!="",j="foo"-8                25773261      38777231      +50.46%
BenchmarkPostingsForMatchers/Block/n="1",i=~".+",j="foo"-8              108811784     120685377     +10.91%
BenchmarkPostingsForMatchers/Block/n="1",i=~"1.+",j="foo"-8             20732880      22044406      +6.33%
BenchmarkPostingsForMatchers/Block/n="1",i=~".+",i!="2",j="foo"-8       107433837     124392905     +15.79%
BenchmarkPostingsForMatchers/Block/n="1",i=~".+",i!~"2.*",j="foo"-8     130279842     144817377     +11.16%

benchmark                                                               old allocs     new allocs     delta
BenchmarkPostingsForMatchers/Head/n="1"-8                               6              6              +0.00%
BenchmarkPostingsForMatchers/Head/n="1",j="foo"-8                       11             11             +0.00%
BenchmarkPostingsForMatchers/Head/j="foo",n="1"-8                       11             11             +0.00%
BenchmarkPostingsForMatchers/Head/n="1",j!="foo"-8                      13             13             +0.00%
BenchmarkPostingsForMatchers/Head/i=~".*"-8                             10             9              -10.00%
BenchmarkPostingsForMatchers/Head/i=~".+"-8                             100039         100039         +0.00%
BenchmarkPostingsForMatchers/Head/i=~""-8                               100043         100043         +0.00%
BenchmarkPostingsForMatchers/Head/i!=""-8                               100039         100039         +0.00%
BenchmarkPostingsForMatchers/Head/n="1",i=~".*",j="foo"-8               15             14             -6.67%
BenchmarkPostingsForMatchers/Head/n="1",i=~".*",i!="2",j="foo"-8        21             20             -4.76%
BenchmarkPostingsForMatchers/Head/n="1",i!=""-8                         100044         100044         +0.00%
BenchmarkPostingsForMatchers/Head/n="1",i!="",j="foo"-8                 100048         100048         +0.00%
BenchmarkPostingsForMatchers/Head/n="1",i=~".+",j="foo"-8               100048         100048         +0.00%
BenchmarkPostingsForMatchers/Head/n="1",i=~"1.+",j="foo"-8              11149          11149          +0.00%
BenchmarkPostingsForMatchers/Head/n="1",i=~".+",i!="2",j="foo"-8        100054         100054         +0.00%
BenchmarkPostingsForMatchers/Head/n="1",i=~".+",i!~"2.*",j="foo"-8      111236         111236         +0.00%
BenchmarkPostingsForMatchers/Block/n="1"-8                              6              6              +0.00%
BenchmarkPostingsForMatchers/Block/n="1",j="foo"-8                      11             11             +0.00%
BenchmarkPostingsForMatchers/Block/j="foo",n="1"-8                      11             11             +0.00%
BenchmarkPostingsForMatchers/Block/n="1",j!="foo"-8                     13             14             +7.69%
BenchmarkPostingsForMatchers/Block/i=~".*"-8                            10             10             +0.00%
BenchmarkPostingsForMatchers/Block/i=~".+"-8                            100039         100040         +0.00%
BenchmarkPostingsForMatchers/Block/i=~""-8                              100043         100044         +0.00%
BenchmarkPostingsForMatchers/Block/i!=""-8                              100039         100040         +0.00%
BenchmarkPostingsForMatchers/Block/n="1",i=~".*",j="foo"-8              15             15             +0.00%
BenchmarkPostingsForMatchers/Block/n="1",i=~".*",i!="2",j="foo"-8       21             21             +0.00%
BenchmarkPostingsForMatchers/Block/n="1",i!=""-8                        100044         100045         +0.00%
BenchmarkPostingsForMatchers/Block/n="1",i!="",j="foo"-8                100048         100049         +0.00%
BenchmarkPostingsForMatchers/Block/n="1",i=~".+",j="foo"-8              100048         100049         +0.00%
BenchmarkPostingsForMatchers/Block/n="1",i=~"1.+",j="foo"-8             11149          11150          +0.01%
BenchmarkPostingsForMatchers/Block/n="1",i=~".+",i!="2",j="foo"-8       100054         100055         +0.00%
BenchmarkPostingsForMatchers/Block/n="1",i=~".+",i!~"2.*",j="foo"-8     111236         111238         +0.00%

benchmark                                                               old bytes     new bytes     delta
BenchmarkPostingsForMatchers/Head/n="1"-8                               296           296           +0.00%
BenchmarkPostingsForMatchers/Head/n="1",j="foo"-8                       424           424           +0.00%
BenchmarkPostingsForMatchers/Head/j="foo",n="1"-8                       424           424           +0.00%
BenchmarkPostingsForMatchers/Head/n="1",j!="foo"-8                      488           488           +0.00%
BenchmarkPostingsForMatchers/Head/i=~".*"-8                             1606154       1606046       -0.01%
BenchmarkPostingsForMatchers/Head/i=~".+"-8                             17264772      17264788      +0.00%
BenchmarkPostingsForMatchers/Head/i=~""-8                               17264822      17264822      +0.00%
BenchmarkPostingsForMatchers/Head/i!=""-8                               17264616      17264616      +0.00%
BenchmarkPostingsForMatchers/Head/n="1",i=~".*",j="foo"-8               1606282       1606200       -0.01%
BenchmarkPostingsForMatchers/Head/n="1",i=~".*",i!="2",j="foo"-8        1606460       1606392       -0.00%
BenchmarkPostingsForMatchers/Head/n="1",i!=""-8                         17264744      17264744      +0.00%
BenchmarkPostingsForMatchers/Head/n="1",i!="",j="foo"-8                 17264872      17264872      +0.00%
BenchmarkPostingsForMatchers/Head/n="1",i=~".+",j="foo"-8               17265044      17265044      +0.00%
BenchmarkPostingsForMatchers/Head/n="1",i=~"1.+",j="foo"-8              3148310       3148256       -0.00%
BenchmarkPostingsForMatchers/Head/n="1",i=~".+",i!="2",j="foo"-8        17265220      17265204      -0.00%
BenchmarkPostingsForMatchers/Head/n="1",i=~".+",i!~"2.*",j="foo"-8      20417074      20417036      -0.00%
BenchmarkPostingsForMatchers/Block/n="1"-8                              296           296           +0.00%
BenchmarkPostingsForMatchers/Block/n="1",j="foo"-8                      424           424           +0.00%
BenchmarkPostingsForMatchers/Block/j="foo",n="1"-8                      424           424           +0.00%
BenchmarkPostingsForMatchers/Block/n="1",j!="foo"-8                     1464          1496          +2.19%
BenchmarkPostingsForMatchers/Block/i=~".*"-8                            1606070       1606062       -0.00%
BenchmarkPostingsForMatchers/Block/i=~".+"-8                            17264696      17264731      +0.00%
BenchmarkPostingsForMatchers/Block/i=~""-8                              17264713      17264763      +0.00%
BenchmarkPostingsForMatchers/Block/i!=""-8                              17264600      17264632      +0.00%
BenchmarkPostingsForMatchers/Block/n="1",i=~".*",j="foo"-8              1606191       1606198       +0.00%
BenchmarkPostingsForMatchers/Block/n="1",i=~".*",i!="2",j="foo"-8       1606390       1606390       +0.00%
BenchmarkPostingsForMatchers/Block/n="1",i!=""-8                        17264728      17264760      +0.00%
BenchmarkPostingsForMatchers/Block/n="1",i!="",j="foo"-8                17264856      17264888      +0.00%
BenchmarkPostingsForMatchers/Block/n="1",i=~".+",j="foo"-8              17264952      17264998      +0.00%
BenchmarkPostingsForMatchers/Block/n="1",i=~"1.+",j="foo"-8             3148222       3148255       +0.00%
BenchmarkPostingsForMatchers/Block/n="1",i=~".+",i!="2",j="foo"-8       17265130      17265174      +0.00%
BenchmarkPostingsForMatchers/Block/n="1",i=~".+",i!~"2.*",j="foo"-8     20416901      20416985      +0.00%
```

Run 2:
```
benchmark                                                               old ns/op     new ns/op     delta
BenchmarkPostingsForMatchers/Head/n="1"-8                               838           822           -1.91%
BenchmarkPostingsForMatchers/Head/n="1",j="foo"-8                       1103          1091          -1.09%
BenchmarkPostingsForMatchers/Head/j="foo",n="1"-8                       1108          1130          +1.99%
BenchmarkPostingsForMatchers/Head/n="1",j!="foo"-8                      1300          1307          +0.54%
BenchmarkPostingsForMatchers/Head/i=~".*"-8                             159238038     100172836     -37.09%
BenchmarkPostingsForMatchers/Head/i=~".+"-8                             194661825     199747583     +2.61%
BenchmarkPostingsForMatchers/Head/i=~""-8                               116276898     118715260     +2.10%
BenchmarkPostingsForMatchers/Head/i!=""-8                               95561357      98064492      +2.62%
BenchmarkPostingsForMatchers/Head/n="1",i=~".*",j="foo"-8               162393387     104570561     -35.61%
BenchmarkPostingsForMatchers/Head/n="1",i=~".*",i!="2",j="foo"-8        161245031     100461299     -37.70%
BenchmarkPostingsForMatchers/Head/n="1",i!=""-8                         97333957      97299090      -0.04%
BenchmarkPostingsForMatchers/Head/n="1",i!="",j="foo"-8                 97928224      97078094      -0.87%
BenchmarkPostingsForMatchers/Head/n="1",i=~".+",j="foo"-8               195668253     201641924     +3.05%
BenchmarkPostingsForMatchers/Head/n="1",i=~"1.+",j="foo"-8              99083032      40633715      -58.99%
BenchmarkPostingsForMatchers/Head/n="1",i=~".+",i!="2",j="foo"-8        192930685     198022182     +2.64%
BenchmarkPostingsForMatchers/Head/n="1",i=~".+",i!~"2.*",j="foo"-8      291180238     241260442     -17.14%
BenchmarkPostingsForMatchers/Block/n="1"-8                              32137         31918         -0.68%
BenchmarkPostingsForMatchers/Block/n="1",j="foo"-8                      571448        575307        +0.68%
BenchmarkPostingsForMatchers/Block/j="foo",n="1"-8                      561969        624371        +11.10%
BenchmarkPostingsForMatchers/Block/n="1",j!="foo"-8                     537670        598385        +11.29%
BenchmarkPostingsForMatchers/Block/i=~".*"-8                            87621402      86410014      -1.38%
BenchmarkPostingsForMatchers/Block/i=~".+"-8                            105747874     118690845     +12.24%
BenchmarkPostingsForMatchers/Block/i=~""-8                              30780658      43424744      +41.08%
BenchmarkPostingsForMatchers/Block/i!=""-8                              24190838      36082068      +49.16%
BenchmarkPostingsForMatchers/Block/n="1",i=~".*",j="foo"-8              88831923      86343732      -2.80%
BenchmarkPostingsForMatchers/Block/n="1",i=~".*",i!="2",j="foo"-8       90105982      87924914      -2.42%
BenchmarkPostingsForMatchers/Block/n="1",i!=""-8                        24167958      38334042      +58.62%
BenchmarkPostingsForMatchers/Block/n="1",i!="",j="foo"-8                25283600      37716089      +49.17%
BenchmarkPostingsForMatchers/Block/n="1",i=~".+",j="foo"-8              106877330     118646879     +11.01%
BenchmarkPostingsForMatchers/Block/n="1",i=~"1.+",j="foo"-8             20695849      21726110      +4.98%
BenchmarkPostingsForMatchers/Block/n="1",i=~".+",i!="2",j="foo"-8       106548005     118119175     +10.86%
BenchmarkPostingsForMatchers/Block/n="1",i=~".+",i!~"2.*",j="foo"-8     130575173     144969529     +11.02%

benchmark                                                               old allocs     new allocs     delta
BenchmarkPostingsForMatchers/Head/n="1"-8                               6              6              +0.00%
BenchmarkPostingsForMatchers/Head/n="1",j="foo"-8                       11             11             +0.00%
BenchmarkPostingsForMatchers/Head/j="foo",n="1"-8                       11             11             +0.00%
BenchmarkPostingsForMatchers/Head/n="1",j!="foo"-8                      13             13             +0.00%
BenchmarkPostingsForMatchers/Head/i=~".*"-8                             10             9              -10.00%
BenchmarkPostingsForMatchers/Head/i=~".+"-8                             100039         100039         +0.00%
BenchmarkPostingsForMatchers/Head/i=~""-8                               100043         100043         +0.00%
BenchmarkPostingsForMatchers/Head/i!=""-8                               100039         100039         +0.00%
BenchmarkPostingsForMatchers/Head/n="1",i=~".*",j="foo"-8               15             14             -6.67%
BenchmarkPostingsForMatchers/Head/n="1",i=~".*",i!="2",j="foo"-8        21             20             -4.76%
BenchmarkPostingsForMatchers/Head/n="1",i!=""-8                         100044         100044         +0.00%
BenchmarkPostingsForMatchers/Head/n="1",i!="",j="foo"-8                 100048         100048         +0.00%
BenchmarkPostingsForMatchers/Head/n="1",i=~".+",j="foo"-8               100048         100048         +0.00%
BenchmarkPostingsForMatchers/Head/n="1",i=~"1.+",j="foo"-8              11149          11149          +0.00%
BenchmarkPostingsForMatchers/Head/n="1",i=~".+",i!="2",j="foo"-8        100054         100054         +0.00%
BenchmarkPostingsForMatchers/Head/n="1",i=~".+",i!~"2.*",j="foo"-8      111236         111236         +0.00%
BenchmarkPostingsForMatchers/Block/n="1"-8                              6              6              +0.00%
BenchmarkPostingsForMatchers/Block/n="1",j="foo"-8                      11             11             +0.00%
BenchmarkPostingsForMatchers/Block/j="foo",n="1"-8                      11             11             +0.00%
BenchmarkPostingsForMatchers/Block/n="1",j!="foo"-8                     13             14             +7.69%
BenchmarkPostingsForMatchers/Block/i=~".*"-8                            10             10             +0.00%
BenchmarkPostingsForMatchers/Block/i=~".+"-8                            100039         100040         +0.00%
BenchmarkPostingsForMatchers/Block/i=~""-8                              100043         100044         +0.00%
BenchmarkPostingsForMatchers/Block/i!=""-8                              100039         100040         +0.00%
BenchmarkPostingsForMatchers/Block/n="1",i=~".*",j="foo"-8              15             15             +0.00%
BenchmarkPostingsForMatchers/Block/n="1",i=~".*",i!="2",j="foo"-8       21             21             +0.00%
BenchmarkPostingsForMatchers/Block/n="1",i!=""-8                        100044         100045         +0.00%
BenchmarkPostingsForMatchers/Block/n="1",i!="",j="foo"-8                100048         100049         +0.00%
BenchmarkPostingsForMatchers/Block/n="1",i=~".+",j="foo"-8              100048         100049         +0.00%
BenchmarkPostingsForMatchers/Block/n="1",i=~"1.+",j="foo"-8             11149          11150          +0.01%
BenchmarkPostingsForMatchers/Block/n="1",i=~".+",i!="2",j="foo"-8       100054         100055         +0.00%
BenchmarkPostingsForMatchers/Block/n="1",i=~".+",i!~"2.*",j="foo"-8     111236         111238         +0.00%

benchmark                                                               old bytes     new bytes     delta
BenchmarkPostingsForMatchers/Head/n="1"-8                               296           296           +0.00%
BenchmarkPostingsForMatchers/Head/n="1",j="foo"-8                       424           424           +0.00%
BenchmarkPostingsForMatchers/Head/j="foo",n="1"-8                       424           424           +0.00%
BenchmarkPostingsForMatchers/Head/n="1",j!="foo"-8                      488           488           +0.00%
BenchmarkPostingsForMatchers/Head/i=~".*"-8                             1606168       1606072       -0.01%
BenchmarkPostingsForMatchers/Head/i=~".+"-8                             17264788      17264788      +0.00%
BenchmarkPostingsForMatchers/Head/i=~""-8                               17264799      17264822      +0.00%
BenchmarkPostingsForMatchers/Head/i!=""-8                               17264616      17264616      +0.00%
BenchmarkPostingsForMatchers/Head/n="1",i=~".*",j="foo"-8               1606282       1606200       -0.01%
BenchmarkPostingsForMatchers/Head/n="1",i=~".*",i!="2",j="foo"-8        1606460       1606392       -0.00%
BenchmarkPostingsForMatchers/Head/n="1",i!=""-8                         17264744      17264744      +0.00%
BenchmarkPostingsForMatchers/Head/n="1",i!="",j="foo"-8                 17264872      17264872      +0.00%
BenchmarkPostingsForMatchers/Head/n="1",i=~".+",j="foo"-8               17265028      17265044      +0.00%
BenchmarkPostingsForMatchers/Head/n="1",i=~"1.+",j="foo"-8              3148302       3148256       -0.00%
BenchmarkPostingsForMatchers/Head/n="1",i=~".+",i!="2",j="foo"-8        17265204      17265220      +0.00%
BenchmarkPostingsForMatchers/Head/n="1",i=~".+",i!~"2.*",j="foo"-8      20417050      20417017      -0.00%
BenchmarkPostingsForMatchers/Block/n="1"-8                              296           296           +0.00%
…